refactor: 당첨 결과를 구하는 책임 LottoTicket으로 전가
