+++ src/test/java/domain/WinningResultTest.java
@@ -1,6 +1,6 @@
 package domain;
 
-import static org.assertj.core.api.AssertionsForClassTypes.*;
+import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -10,132 +10,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class WinningResultTest {
-
-	@Test
-	@DisplayName("1등 당첨 결과 확인")
-	void checkFirstWinningResult() {
-		// given
-		Lotto myLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 35)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Lotto winningLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 35)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Number bonusNumber = new Number(36);
-		WinningNumbers winningNumbers = new WinningNumbers(winningLotto, bonusNumber);
-		LottoTicket lottoTicket = new LottoTicket(List.of(myLotto));
-
-		// when
-		WinningResult winningResult = WinningResult.createWinningResult(lottoTicket, winningNumbers);
-
-		// then
-		assertThat(winningResult.getWinningResult().get(LottoRank.FIRST)).isEqualTo(1);
-	}
-
-	@Test
-	@DisplayName("2등 당첨 결과 확인")
-	void checkSecondWinningResult() {
-		// given
-		Lotto myLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 35)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Lotto winningLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 34)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Number bonusNumber = new Number(35);
-		WinningNumbers winningNumbers = new WinningNumbers(winningLotto, bonusNumber);
-		LottoTicket lottoTicket = new LottoTicket(List.of(myLotto));
-
-		// when
-		WinningResult winningResult = WinningResult.createWinningResult(lottoTicket, winningNumbers);
-
-		// then
-		assertThat(winningResult.getWinningResult().get(LottoRank.SECOND)).isEqualTo(1);
-	}
-
-	@Test
-	@DisplayName("3등 당첨 결과 확인")
-	void checkThirdWinningResult() {
-		// given
-		Lotto myLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 35)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Lotto winningLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 34)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Number bonusNumber = new Number(39);
-		WinningNumbers winningNumbers = new WinningNumbers(winningLotto, bonusNumber);
-		LottoTicket lottoTicket = new LottoTicket(List.of(myLotto));
-
-		// when
-		WinningResult winningResult = WinningResult.createWinningResult(lottoTicket, winningNumbers);
-
-		// then
-		assertThat(winningResult.getWinningResult().get(LottoRank.THIRD)).isEqualTo(1);
-	}
-
-	@Test
-	@DisplayName("4등 당첨 결과 확인")
-	void checkFourthWinningResult() {
-		// given
-		Lotto myLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 35)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Lotto winningLotto = new Lotto(Stream.of(1, 5, 9, 11, 17, 34)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Number bonusNumber = new Number(39);
-		WinningNumbers winningNumbers = new WinningNumbers(winningLotto, bonusNumber);
-		LottoTicket lottoTicket = new LottoTicket(List.of(myLotto));
-
-		// when
-		WinningResult winningResult = WinningResult.createWinningResult(lottoTicket, winningNumbers);
-
-		// then
-		assertThat(winningResult.getWinningResult().get(LottoRank.FOURTH)).isEqualTo(1);
-	}
-
-	@Test
-	@DisplayName("5등 당첨 결과 확인")
-	void checkFIFTHWinningResult() {
-		// given
-		Lotto myLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 35)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Lotto winningLotto = new Lotto(Stream.of(1, 5, 9, 12, 17, 34)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Number bonusNumber = new Number(39);
-		WinningNumbers winningNumbers = new WinningNumbers(winningLotto, bonusNumber);
-		LottoTicket lottoTicket = new LottoTicket(List.of(myLotto));
-
-		// when
-		WinningResult winningResult = WinningResult.createWinningResult(lottoTicket, winningNumbers);
-
-		// then
-		assertThat(winningResult.getWinningResult().get(LottoRank.FIFTH)).isEqualTo(1);
-	}
-
-	@Test
-	@DisplayName("1000원으로 5등 당첨 시, 수익률 확인")
-	void checkRateOfProfit() {
-		Lotto myLotto = new Lotto(Stream.of(1, 5, 9, 11, 16, 35)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Lotto winningLotto = new Lotto(Stream.of(1, 5, 9, 22, 34, 44)
-			.map(Number::new)
-			.collect(Collectors.toList()));
-		Number bonusNumber = new Number(39);
-		WinningNumbers winningNumbers = new WinningNumbers(winningLotto, bonusNumber);
-		LottoTicket lottoTicket = new LottoTicket(List.of(myLotto));
-		Money money = new Money(1000);
-
-		// when
-		WinningResult winningResult = WinningResult.createWinningResult(lottoTicket, winningNumbers);
-
-		//then
-		assertThat(winningResult.getRateOfProfit(money)).isEqualTo(5.0);
-	}
 
 	@Test
 	@DisplayName("1000원으로 1등 당첨 시, 수익률 확인")
@@ -152,7 +26,7 @@ class WinningResultTest {
 		Money money = new Money(1000);
 
 		// when
-		WinningResult winningResult = WinningResult.createWinningResult(lottoTicket, winningNumbers);
+		WinningResult winningResult = new WinningResult(lottoTicket.checkLottoTicketWinningResult(winningNumbers));
 
 		//then
 		assertThat(winningResult.getRateOfProfit(money)).isEqualTo(2000000.0);
